# api/wellness-buddy-openapi.yaml
openapi: 3.0.3
info:
  title: Wellness Buddy API
  version: 1.0.0
  description: |
    REST API for the Wellness Buddy app. The client must never access the
    database directly; all reads and writes go through this API layer.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /profiles:
    get:
      tags: [Profiles]
      summary: List profiles
      description: Returns all profiles or filters by role.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [traveler, moderator, admin]
          required: false
          description: Filter by role.
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
    post:
      tags: [Profiles]
      summary: Create a profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
  /profiles/{id}:
    get:
      tags: [Profiles]
      summary: Get a profile by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Profiles]
      summary: Update a profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Profiles]
      summary: Delete a profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '204':
          description: Profile deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProfileId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [traveler, moderator, admin]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProfileCreate:
      type: object
      required: [id, full_name, email]
      properties:
        id:
          type: string
          format: uuid
          description: Must match auth.users.id
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [traveler, moderator, admin]
          default: traveler
    ProfileUpdate:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [traveler, moderator, admin]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true